Extract duplicate route formatting functions to shared module
